# Improve performance of confocal image reconstruction