--- a/lumicks/pylake/scan.py
+++ b/lumicks/pylake/scan.py
@@ -34,8 +34,11 @@ class Scan(Kymo):
         return self.json["scan volume"]["scan axes"][1]["num of pixels"]
 
     def _image(self, color):
-        return reconstruct_image(getattr(self, f"{color}_photon_count").data, self.infowave.data,
-                                 self.pixels_per_line, self.lines_per_frame)
+        if color not in self._cache:
+            photon_counts = getattr(self, f"{color}_photon_count").data
+            self._cache[color] = reconstruct_image(photon_counts, self.infowave.data,
+                                                   self.pixels_per_line, self.lines_per_frame)
+        return self._cache[color]
 
     def _timestamps(self, sample_timestamps):
         return reconstruct_image(sample_timestamps, self.infowave.data, self.pixels_per_line,