Should mostly resolve NG-1472.

Benchmarked with a 2 GiB HDF5 file containing a continuous scan with 212 frames:
```
|            | before (ms) | after (ms) | speedup |
|------------|-------------|------------|---------|
| load       |   1353.3    |     1.1    |    ∞    |
| rgb image  |   3906.6    |   863.9    |   4.5x  |
| num_frames |   0.003     |    19.7    |    ∞    |
|------------|-------------|------------|---------|
|  total     |   5259.9    |   884.7    |  5.95x  |
```

* `load` is just listing the scan using `file.scans`. This was previously very slow for continuous scans because the the number of frames needed to be calculated and this was being done in the constructor. This is now done lazily, but the calculation is also faster. See below.
* `rgb image` indicates the time needed to reconstruct the full RGB image with all the frames. This got a massive speedup thanks to a more efficient way of slicing continuous channels and a faster way of calculating the number of samples per pixel.
* `num_frames` is the time needed to look up the number of frames in a continuous scan. This was previously instant becaues it was calculated up front. This is now done lazily and it uses a dedicated algorithms which only looks at the infowave instead of reconstructing the entire image to get number of frames.

In total, the speedup is quite significant and should also reduce RAM usage.