@onnodb 2019-03-13
Might be good to contact Filip in Delft directly with this change, and see whether it helps him. 

I'll have a closer look later, but I did not this is not going to support 3D scans. That's totally fine (Bluelake itself doesn't support it either), but would it make sense to make a note of that somewhere? Or is there already a check for this somewhere?