@dean0x7d 2019-03-13
> I'll have a closer look later, but I did not this is not going to support 3D scans. That's totally fine (Bluelake itself doesn't support it either), but would it make sense to make a note of that somewhere? Or is there already a check for this somewhere?

I just added an explicit check in `__init__()`. See the last commit.